Refactors code so that circular and square neighbourhood probability classes are in separate modules.
