+++ lib/improver/tests/nbhood/nbhood/test_CircularNeighbourhood.py
@@ -37,7 +37,7 @@ from iris.cube import Cube
 from iris.tests import IrisTest
 import numpy as np
 
-from improver.nbhood.nbhood import CircularNeighbourhood
+from improver.nbhood.circular_kernel import CircularNeighbourhood
 from improver.tests.nbhood.nbhood.test_NeighbourhoodProcessing import (
     SINGLE_POINT_RANGE_2_CENTROID_FLAT, SINGLE_POINT_RANGE_3_CENTROID,
     SINGLE_POINT_RANGE_5_CENTROID, set_up_cube, set_up_cube_lat_long)
